#1: Refactors the time package. "Plain Slot Date" is now "Abstract Slot Date" and "Full Slot Date" is now "Concrete Slot Date" to make the distinction clearer. An abstract date has only the information of epoch and slot number, while the concrete date has the time settings information of a blockchain, and hence, exists in the time system of this particular blockchain.
Additionally, the epoch and slot number are now usigned integer of bit size 64 instead
of big integers of the math/big package. Due to the nature of epoch/slots, math/big is
only needed for computation on the date, but not for the date itself. However, this
could be easily changed in the future without violating the current interface.

Moreover, the code quality has been improved.